Hack to get Livejournal session cookies to work.
I suspect that LWP is stricter than wget in matching cookies
to domains, and that's why wget works with Livejournal cookies,
because Livejournal always puts a '.' in front of the domain
information for its cookies, even when it doesn't need one,
such as the cookies for individual journals. Thus, LWP doesn't
send cookies for individual journals - which are exactly the
session cookies one needs to log in, thus LWP always fails
to log in to Livejournal with cookies.  Other sites, such
as Dreamwidth, give more sensible values for the domain field
of their cookies, so LWP is able to match them correctly.